+++ src/constants/language.constant.ts
@@ -1,0 +1,5 @@
+export const supportedLanguages = [
+    'en',
+    'en-US',
+    'pl'
+] as const;