+++ src/modules/recipe/recipe.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { LoggerService } from '../logger/logger.service';
-import { RecipeRepository } from '../../mongodb/repositories/recipe.repository';
+import { DishRecipeRepository } from '../../mongodb/repositories/dish-recipe.repository';
 import { CreateRecipeDto } from './recipe.dto';
 import { DishRepository } from '../../mongodb/repositories/dish.repository';
-import { RecipeDocument } from '../../mongodb/documents/recipe.document';
+import { DishRecipeDocument } from '../../mongodb/documents/dish-recipe-document';
 import { NotFoundException } from '../../exceptions/not-found.exception';
 import { ContextString, Language } from '../../common/types';
 import { ForbiddenException } from '../../exceptions/forbidden-exception';
@@ -21,13 +21,13 @@ export class RecipeService {
 
     constructor(
         private readonly externalApiService: ExternalApiService,
-        private readonly recipeRepository: RecipeRepository,
+        private readonly recipeRepository: DishRecipeRepository,
         private readonly dishRepository: DishRepository,
         private readonly loggerService: LoggerService,
         private readonly redisService: RedisService
     ) {}
 
-    async create(dishId: string, createRecipeDto: CreateRecipeDto, user: UserAccessTokenPayload): Promise<RecipeDocument> {
+    async create(dishId: string, createRecipeDto: CreateRecipeDto, user: UserAccessTokenPayload): Promise<DishRecipeDocument> {
         const context: ContextString = 'RecipeService/create';
 
         // NOTE: This dish can be unconfirmed because you add dish and recipe at once.