+++ src/modules/user/user.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
 import { MongooseModule } from '@nestjs/mongoose';
-import { USER_ACTION_MODEL, USER_MODEL } from '../../constants/models.constant';
+import { userModel, userActionModel } from '../../common/definitions/mongoose-model.definitions';
 import { JwtManagerModule } from '../jwt-manager/jwt-manager.module';
 import { RedisModule } from '../redis/redis.module';
 import { MailManagerModule } from '../mail-manager/mail-manager.module';
@@ -13,7 +13,7 @@ import { PasswordManagerModule } from '../password-manager/password-manager.modu
 
 @Module({
     imports: [
-        MongooseModule.forFeature([USER_MODEL, USER_ACTION_MODEL]),
+        MongooseModule.forFeature([userModel, userActionModel]),
         JwtManagerModule,
         RedisModule,
         MailManagerModule,