+++ src/mongodb/documents/user-search-query.document.ts
@@ -1,6 +1,6 @@
 import { Document } from 'mongoose';
 
-export interface SearchQueryDocument extends Document {
+export interface UserSearchQueryDocument extends Document {
     readonly ingredients: string[];
     readonly date: Date;
     readonly login: string;