+++ src/mongodb/schemas/search-query.schema.ts
@@ -1,7 +1,0 @@
-import * as mongoose from 'mongoose';
-
-export const SearchQuerySchema = new mongoose.Schema({
-    ingredients: [String],
-    date: Date,
-    login: String
-});