+++ src/mongodb/schemas/user-search-query.schema.ts
@@ -1,0 +1,9 @@
+import * as mongoose from 'mongoose';
+
+export const UserSearchQuerySchema = new mongoose.Schema({
+    ingredients: [String],
+    date: Date,
+    login: String
+});
+
+UserSearchQuerySchema.index({ login: 1 });